+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Dev
+
+* Replaced use of M2Crypto with Cryptography library
+* Added python3 support
+
 ## 0.1.0
 
 * Added a timeout parameter to verify() and get_ticket_data(). Defaults